+++ src/ui/CategoryManger.kt
@@ -1,4 +1,4 @@
-package UI
+package ui
 
 object CategoryManger {
