+++ src/ui/ExpenseManager.kt
@@ -1,75 +1,68 @@
-package UI
+package ui
 
-object IncomeManager {
+object ExpenseManager {
 
 
-
-    fun addIncomeMenu() {
-
+    fun addExpenseMenu() {
         print("---------------------------\n" +
-                "Lets add income\n" +
+                "Lets add Expense\n" +
                 "Enter Amount: ")
-        var incomeAmount: Double = readln().toDouble()
+        var expenseAmount: Double = readln().toDouble()
+        println("your transaction added successfully ")
+        print("Chose Expense Category\n")
 
-        print("Chose Income Category\n")
-
-        // print Income Category
-
-        // pass to add income function
-        println("Income added successfully")
-
+        // print Expense Category
+        // pass to add Expense function
     }
 
 
-    fun viewIncomeTransaction() {
+    fun viewExpenseTransaction() {
         print("---------------------------\n" +
-                "Here is all the income\n")
-        // call view income function
+                "Here is all the expense\n")
+        // call view expense function
         print("What do you want to do?\n" +
                 "1. Edit\n" + // done
                 "2. Delete\n" + // done
                 "3. Back\n" +
                 "Enter Your option: ")
         when(readln().toIntOrNull()){
-            1 -> editIncomeTransaction()
-            2 -> deleteIncomeTransaction()
+            1 -> editExpenseTransaction()
+            2 -> deleteExpenseTransaction()
             3 -> return
             null->println("Invalid Input try again")
-            else -> println("Enter a valid number between 1 - 5")
+            else -> println("Enter a valid number between 1 - 3")
         }
     }
 
 
-    fun deleteIncomeTransaction() {
+    fun deleteExpenseTransaction() {
         print("Enter Transaction ID: ")
-        var deleteIncomeTransactionId: Int = readln().toInt()
+        var deleteExpenseTransactionId: Int = readln().toInt()
         // this Transaction exist ??
         // call delete fun
     }
 
-    fun editIncomeTransaction() {
-        var editIncomeTransactionAmount: Double
+    fun editExpenseTransaction() {
+        var editExpenseTransactionAmount: Double
         print("Enter Transaction ID: ")
-        var editIncomeTransactionID: Int = readln().toInt()
+        var editExpenseTransactionID: Int = readln().toInt()
         print("What do you want to edit (1. amount 2. category): ")
         when(readln().toIntOrNull()){
             1 -> {
                 print("Enter New Amount: ")
-                editIncomeTransactionAmount = readln().toDouble()
+                editExpenseTransactionAmount = readln().toDouble()
                 // pass to function and update
                 println("updated")
             }
             2 -> {
-                print("Chose income Category\n")
-                // view income category
+                print("Chose expense Category\n")
+                // view expense category
                 println("updated")
             }
             null->println("Invalid Input try again")
             else -> println("Enter a valid number 1 or 2")
         }
     }
-
-
 
 
 }
