+++ src/ui/FinanceCLI.kt
@@ -1,4 +1,4 @@
-package UI
+package ui
 
 fun runApp(){
     MenusManager.startMenu()
