+++ src/ui/IncomeManager.kt
@@ -1,68 +1,75 @@
-package UI
+package ui
 
-object ExpenseManager {
+object IncomeManager {
 
 
-    fun addExpenseMenu() {
+
+    fun addIncomeMenu() {
+
         print("---------------------------\n" +
-                "Lets add Expense\n" +
+                "Lets add income\n" +
                 "Enter Amount: ")
-        var expenseAmount: Double = readln().toDouble()
-        println("your transaction added successfully ")
-        print("Chose Expense Category\n")
+        var incomeAmount: Double = readln().toDouble()
 
-        // print Expense Category
-        // pass to add Expense function
+        print("Chose Income Category\n")
+
+        // print Income Category
+
+        // pass to add income function
+        println("Income added successfully")
+
     }
 
 
-    fun viewExpenseTransaction() {
+    fun viewIncomeTransaction() {
         print("---------------------------\n" +
-                "Here is all the expense\n")
-        // call view expense function
+                "Here is all the income\n")
+        // call view income function
         print("What do you want to do?\n" +
                 "1. Edit\n" + // done
                 "2. Delete\n" + // done
                 "3. Back\n" +
                 "Enter Your option: ")
         when(readln().toIntOrNull()){
-            1 -> editExpenseTransaction()
-            2 -> deleteExpenseTransaction()
+            1 -> editIncomeTransaction()
+            2 -> deleteIncomeTransaction()
             3 -> return
             null->println("Invalid Input try again")
-            else -> println("Enter a valid number between 1 - 3")
+            else -> println("Enter a valid number between 1 - 5")
         }
     }
 
 
-    fun deleteExpenseTransaction() {
+    fun deleteIncomeTransaction() {
         print("Enter Transaction ID: ")
-        var deleteExpenseTransactionId: Int = readln().toInt()
+        var deleteIncomeTransactionId: Int = readln().toInt()
         // this Transaction exist ??
         // call delete fun
     }
 
-    fun editExpenseTransaction() {
-        var editExpenseTransactionAmount: Double
+    fun editIncomeTransaction() {
+        var editIncomeTransactionAmount: Double
         print("Enter Transaction ID: ")
-        var editExpenseTransactionID: Int = readln().toInt()
+        var editIncomeTransactionID: Int = readln().toInt()
         print("What do you want to edit (1. amount 2. category): ")
         when(readln().toIntOrNull()){
             1 -> {
                 print("Enter New Amount: ")
-                editExpenseTransactionAmount = readln().toDouble()
+                editIncomeTransactionAmount = readln().toDouble()
                 // pass to function and update
                 println("updated")
             }
             2 -> {
-                print("Chose expense Category\n")
-                // view expense category
+                print("Chose income Category\n")
+                // view income category
                 println("updated")
             }
             null->println("Invalid Input try again")
             else -> println("Enter a valid number 1 or 2")
         }
     }
+
+
 
 
 }
