+++ src/ui/MenusManager.kt
@@ -1,4 +1,4 @@
-package UI
+package ui
 
 object MenusManager {
